New examples I created that work with Blynk.
Button_Switch_LED.ino - Features the ability to control an LED through the form of a switch both on a device connected via WiFi or by a switch connected to the ESP32 module itself.

Intruder_Detection_Analysis.ino - Brings security to the ESP32 module and enables an ultrasonic sensor to not only detect movement but warns the user if an intruder has been detected or if the area is clear.

LED_Control_Blynk.ino - Another code I tried before Button_Switch_LED.ino which is less advanced but should enable the user to control an LED through WiFi on Blynk.

Temperature_Humidity_Sensor_V2.ino - Better than the previous iteration which was V1. This piece of code enables a DHT11 sensor module to display the current temperature of the atmosphere in Celsius and display the humidity as well.

Ultrasonic.ino - Another similar version to Intruder_Detection_Analysis.ino that is less complex but practically does the same thing.

Lastly, Smart_Home.ino - A file that introduces almost all pieces of code up to this point which features the Temperature and Humidity Sensor, LED Control, and the Intruder Detection Analysis all into one file, not yet tested but should be close to working with a bit of tweaking.